Updates to handling of user specific items
It makes the most sense to run this script in the system user space, meaning HKCU doesn't exist technically, as well as the scheduled task would be done under the system user.

Instead updating to set the task to run as any built in user, and having the registry function basically parse out any users, and loop through their notify settings to look for twingate.exe, and update it if it's there.

Tested fine on my machines!