Update twingate_client_installer.ps1
Updates to the client deployment script to include the check for the user being logged in or not.  This is an optional feature that will create a scheduled task that can check, and if the user is logged out (based off of a test) then notify the user via toast.