Initial MacOS Deployment Script
Initial run at the script, includes most of the same features as the Powershell version.  Still meant to be pushed out via MDM but till take on downloading the actual pkg file, installing it, configuring it with the plist, as well as optionally creating a machinekey.conf and a script to force the client to be running even if the user quits.