Update twingate_client_installer.ps1
Adding an auto-update feature, which will save this script into the twingate folder locally, and create a scheduled task to run it every x days.  That way when it's run by the system user, it will repeat this whole process (install client update, re-set scheduled tasks, etc).

Also cleaned up a couple of other things such as how the script triggers the app to start.  Given this is usually run as a system user the client probably doesn't start cleanly for normal users, so now it uses a scheduled task (which is scheduled to never run) to start the app manually.