Adding support for the NextDNS root cert
There are some situations where the DNS filtering block page will not load properly due to HTTPS -> HTTP downgrade, and either the browser extension or the NextDNS root certificate is needed to resolve.

As this is a third party certificate not handled by Twingate, there needs to be a way to both install and uninstall it later on if necessary, so adding it to both scripts.