Adding a detection script
Toying with the idea of moving from a pure fire and forget it style deployment/installation script to something that does detection and remediation, which is supported through Intune at the least.

The included script is an initial effort to detect whether the installed copy of the Twingate client is the same version as the last one listed in the changelog RSS, if not then it will fail in a way that can trigger a remediation/install script.