Adding method to stay behind a version
For both detection and remediation (client installer) scripts, updated the logic to allow the admin to choose to "stay behind" versions.

If you set the flag to a number, such as 1 or 2, then it will look through the changelog and find the next 1 or 2 older versions from the latest.

So 0 is latest, 1 is next to latest, 2 is next-next to latest, and so on.

This allows admins to avoid the latest and greatest Windows Client when doing updates, or situations where they're installing a brand new released today client, and a local EDR may block usage due to the age of the client application.

Important that if you're using both of these scripts in Intune or any other MDM that supports a detect and remediate style deployment, that you set the same versions to stay behind in both before using the scripts together, otherwise you'll end up pulling down the wrong version.